# api.yaml
swagger: "2.0"
info:
  title: Message API
  description: Schedule and send messages to recipients
  version: 1.0.0
host: "localhost:8080"
basePath: /api/
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json


security:
  - ApiKeyAuth: [ ]

securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: Authorization

paths:
  /messages:
    get:
      summary: Get all messages
      operationId: getMessages
      tags:
        - messages
      parameters:
        - $ref: "#/parameters/offset"
        - $ref: "#/parameters/limit"
        - $ref: "#/parameters/order"
        - $ref: "#/parameters/status"
        - $ref: "#/parameters/recipient"
      responses:
        200:
          $ref: "#/responses/MessageList"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        500:
          $ref: "#/responses/InternalServerError"
    post:
      summary: Schedule a new message
      operationId: sendMessage
      tags:
        - messages
      parameters:
        - name: request
          in: body
          required: true
          schema:
            $ref: "#/definitions/CreateMessageRequest"
      responses:
        200:
          $ref: "#/responses/Message"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        500:
          $ref: "#/responses/InternalServerError"
parameters:
  limit:
    name: limit
    in: query
    description: How many results want to list per page
    type: integer
    required: false
    default: 100
    minimum: 1
    maximum: 100
  offset:
    name: offset
    in: query
    description: Query offset
    type: integer
    required: false
    default: 0
  order:
    name: order
    in: query
    type: string
    required: false
  status:
    name: status
    in: query
    type: string
    enum:
      - SCHEDULED
      - SENT
      - FAILED
  recipient:
    name: recipient
    in: query
    type: string

responses:
  NoContent:
    description: Operation successfully
  InternalServerError:
    description: Unknown error has occurred
    schema:
      $ref: "#/definitions/ApiError"
  Unauthorized:
    description: Authorization information is missing, invalid or user does not have permission to take requested action
    schema:
      $ref: "#/definitions/ApiError"
  BadRequest:
    description: Given request format is not same with expected format or validation failed
    schema:
      $ref: "#/definitions/ApiError"
  Message:
    description: Message object
    schema:
      $ref: "#/definitions/Message"
  MessageList:
    description: List of messages
    schema:
      $ref: "#/definitions/MessageList"


definitions:
  PaginatedResponse:
    type: object
    required:
      - offset
      - limit
      - total
    properties:
      offset:
        type: integer
        example: 10
      limit:
        type: integer
        example: 5
      total:
        type: integer
        description: How many total entity exist for given query
        example: 100
  ApiError:
    type: "object"
    properties:
      code:
        type: integer
        format: int64
        example: 1001
        description: Uniq error code
      description:
        type: string
        description: Human readable explanation about error
  CreateMessageRequest:
    type: object
    required:
      - recipient
      - content
    properties:
      recipient:
        type: string
        maxLength: 255
        minLength: 15
      content:
        type: string
        maxLength: 2048
        minLength: 1
  Message:
    type: object
    properties:
      id:
        type: integer
        format: int64
      recipient:
        type: string
      content:
        type: string
      sentTime:
        type: string
        format: date-time
        x-nullable: true
        x-omitempty: false
      status:
        type: string
        enum:
          - SCHEDULED
          - SENT
          - FAILED
      messageID:
        type: string
        x-nullable: true
        x-omitempty: false
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  MessageList:
    allOf:
      - $ref: "#/definitions/PaginatedResponse"
      - type: object
        properties:
          data:
            type: array
            items:
              $ref: "#/definitions/Message"